% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOAPI.R
\name{get.banAPI}
\alias{get.banAPI}
\title{Query locations associated with a requested address in France}
\usage{
get.banAPI(query = "8 bd du port", geo.place = F)
}
\description{
Use French *Base d'Adresses Nationales* (\url{https://adresse.data.gouv.fr/api/})
to get coordinates of an address or a place
}
\examples{
# Search 8 boulevard du Port in all French cities
get.banAPI()$features
# Search 8 boulevard du Port around a point
get.banAPI('8 bd du port&lat=48.789&lon=2.789')$features
# Search for Kebab everywhere
get.banAPI('kebab',T)$features
# Search a specific adress
get.banAPI('36 Quai des Orfèvres')$features
}
