% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOAPI.R
\name{get.GEOAPI.byzip}
\alias{get.GEOAPI.byzip}
\alias{get.GEOAPI}
\title{Query French city informations with GEO API}
\usage{
get.GEOAPI.byzip(ZIPcode)

get.GEOAPI(query,
  fields = "nom,code,codesPostaux,codeDepartement,codeRegion,population")
}
\arguments{
\item{ZIPcode}{ZIP code to query}
}
\value{
List returning all cities that match a request.
 with this ZIP code with some. For instance, by default,
`code` (INSEE city identifier),
`codesPostaux` (ZIP code), `codeDepartement` (department code),
`codeRegion` (region code), `population`
}
\description{
Simple calls to \url{https://api.gouv.fr/api/api-geo.html},
French government statistics (from Etalab). \code{get.GEOAPI.byzip}
returns the cities associated to the ZIP code requested
\code{get.GEOAPI} allows more complex calls
}
\examples{
get.GEOAPI.byzip("35220")
get.GEOAPI.byzip("75001")
get.GEOAPI("nom=versailles")
get.GEOAPI("lon=2.1301&lat=48.8014")
get.GEOAPI("nom=toulouse", fields = "nom, code, codesPostaux, population")
}
